% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{bootstrap_ci}
\alias{bootstrap_ci}
\title{Calculate block bootstrap confidence interval}
\usage{
bootstrap_ci(match, dat, treat_variable = "treat",
  outcome_variable = "y_obs", trajectory_id = "personID",
  ci_level = 0.95, reps = 1000, outcome_model = NA)
}
\arguments{
\item{match}{GroupMatch output}

\item{dat}{Dataset GroupMatch was performed on}

\item{treat_variable}{Name of the treatment variable in the dataset}

\item{outcome_variable}{Name of the outcome variable in the dataset}

\item{trajectory_id}{Name of the column that identifies trajectories}

\item{ci_level}{Confidence level}

\item{reps}{Number of bootstrap reps}

\item{outcome_model}{Outcome model, if bias correction is used}
}
\value{
Vector with confidence interval endpoints
}
\description{
Calculate block bootstrap confidence interval
}
