% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{control_test}
\alias{control_test}
\title{Run falsification test}
\usage{
control_test(control_data, control_variables, outcome_model = NA,
  outcome_variable, time_variable, trajectory_id, time_points,
  rand_reps = 100, caliper = 0.2, reps = 1000)
}
\arguments{
\item{control_data}{Dataset with only control units}

\item{control_variables}{Vector of variables to match on}

\item{outcome_model}{Outcome model, if bias correction is used}

\item{outcome_variable}{Name of the outcome variable in the dataset}

\item{time_variable}{Name of the time variable in the dataset}

\item{trajectory_id}{Name of the column that identifies trajectories}

\item{time_points}{Vector of two time points to run falsification test on}

\item{rand_reps}{Number of reps to randomly divide dataset}

\item{caliper}{Matching caliper}

\item{reps}{Number of permutation test reps}
}
\value{
List with p-value from test and mahalanobis distance of dataset for random split
}
\description{
Run falsification test
}
