% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupmatch.R
\name{groupmatch}
\alias{groupmatch}
\title{Optimal full matching with control groups}
\usage{
groupmatch(
  x,
  group = NULL,
  allow_duplicates = FALSE,
  min.controls = 0,
  max.controls = Inf,
  omit.fraction = NULL,
  mean.controls = NULL,
  tol = 0.001,
  data = NULL,
  ...
)
}
\arguments{
\item{x}{Any valid input to \code{match_on}. \code{groupmatch} will use
\code{x} and any optional arguments to generate a distance before performing
the matching.

If \code{x} is a numeric vector, there must also be passed a vector \code{z}
indicating grouping. Both vectors must be named.

Alternatively, a precomputed distance may be entered. A matrix of
non-negative discrepancies, each indicating the permissibility and
desirability of matching the unit corresponding to its row (a 'treatment') to
the unit corresponding to its column (a 'control'); or, better, a distance
specification as produced by \code{\link{match_on}}.  A simple distance
specification - for example, a matrix of propensity score distances - can be
enhanced by combining it with matrices representing exact-match or other
caliper restrictions.  The final matrix, including these constraints, is a
valid input to \code{groupmatch}.}

\item{group}{Grouping variable for control group.  In the case of rolling
enrollment, this will be a unique subject identifier pertaining to all
'copies' or 'versions' of the same subject.}

\item{allow_duplicates}{When \code{allow_duplicates} is FALSE, the algorithm
ensures that exactly one 'copy' or 'version' of each unique potential comparison
subject is included in the matched comparison group, corresponding to Problem A
in Pimentel et al. (2019).  When \code{allow_duplicates} is TRUE, the algorithm
permits different versions of the same potential comparison subject to match to different
treatment subjects, corresponding to Problem B in Pimentel et al (2019).  So, for
example, when allow_duplicates is FALSE, only one version of unique potential
comparison subject C1 could match to any treatment subject; when allow_duplicates
is TRUE, versions C1a and C1d could match to treatment subjects  T4 and T7, respectively.}

\item{min.controls}{The minimum ratio of controls to treatments that is to
be permitted within a matched set: should be non-negative and finite.  If
\code{min.controls} is not a whole number, the reciprocal of a whole number,
or zero, then it is rounded \emph{down} to the nearest whole number or
reciprocal of a whole number.

Currently, \code{groupmatch} requires that \code{min.controls} be greater than or
equal to 1.  \code{min.controls} less than one implies matching with replacement,
which scenario is currently under development.

When matching within subclasses (such as those created by
\code{\link{exactMatch}}), \code{min.controls} may be a named numeric vector
separately specifying the minimum permissible ratio of controls to treatments
for each subclass.  The names of this vector should include names of all
subproblems \code{distance}.}

\item{max.controls}{The maximum ratio of controls to treatments that is
to be permitted within a matched set: should be positive and numeric.
If \code{max.controls} is not a whole number, the reciprocal of a
whole number, or \code{Inf}, then it is rounded \emph{up} to the
nearest whole number or reciprocal of a whole number.

When matching within subclasses (such as those created by
\code{\link{exactMatch}}), \code{max.controls} may be a named numeric vector
separately specifying the maximum permissible ratio of controls to treatments
in each subclass.}

\item{omit.fraction}{Optionally, specify what fraction of controls or treated
subjects are to be rejected.  If \code{omit.fraction} is a positive fraction
less than one, then \code{groupmatch} leaves up to that fraction of the control
reservoir unmatched.  If \code{omit.fraction} is a negative number greater
than -1, then \code{groupmatch} leaves up to |\code{omit.fraction}| of the
treated group unmatched.  Positive values are only accepted if
\code{max.controls} >= 1; negative values, only if \code{min.controls} <= 1.
If neither \code{omit.fraction} nor \code{mean.controls} is specified, then
only those treated and control subjects without permissible matches among the
control and treated subjects, respectively, are omitted.

When matching within subclasses (such as those created by
\code{\link{exactMatch}}), \code{omit.fraction} specifies the fraction of
controls to be rejected in each subproblem, a parameter that can be made to
differ by subclass by setting \code{omit.fraction} equal to a named numeric
vector of fractions.

At most one of \code{mean.controls} and \code{omit.fraction} can be non-\code{NULL}.}

\item{mean.controls}{Optionally, specify the average number of controls per
treatment to be matched. Must be no less than than \code{min.controls} and no
greater than the either \code{max.controls} or the ratio of total number of
controls versus total number of treated. Some controls will likely not be
matched to ensure meeting this value. If neither \code{omit.fraction} or
\code{mean.controls} are specified, then only those treated and control
subjects without permissible matches among the control and treated subjects,
respectively, are omitted.

When matching within subclasses (such as those created by
\code{\link{exactMatch}}), \code{mean.controls} specifies the average number of
controls per treatment per subproblem, a parameter that can be made to
differ by subclass by setting \code{mean.controls} equal to a named numeric
vector.

At most one of \code{mean.controls} and \code{omit.fraction} can be non-\code{NULL}.}

\item{tol}{Because of internal rounding, \code{groupmatch} may
solve a slightly different matching problem than the one
specified, in which the match generated by
\code{groupmatch} may not coincide with an optimal solution of
the specified problem.  \code{tol} times the number of subjects
to be matched specifies the extent to
which \code{groupmatch}'s output is permitted to differ from an
optimal solution to the original problem, as measured by the
sum of discrepancies for all treatments and controls placed
into the same matched sets.}

\item{data}{Optional \code{data.frame} or \code{vector} to use to get order
of the final matching factor. If a \code{data.frame}, the \code{rownames}
are used. If a vector, the \code{names} are first tried, otherwise the contents
is considered to be a character vector of names. Useful to pass if you want to
combine a match (using, e.g., \code{cbind}) with the data that were used to
generate it (for example, in a propensity score matching).}

\item{...}{Additional arguments, including \code{within}, which may be passed to \code{match_on}.}
}
\value{
A \code{\link{optmatch}} object (\code{factor}) indicating matched groups.
}
\description{
This is an adaption of \code{\link[optmatch]{fullmatch}} to allow for
restrictions when control observations are "grouped". The motivating use
case is when there are multiple observations of control data for each
control subject. In this case, the grouping variable is the subject. We
may want to place restrictions, for example that only one observation of
a subject can be matched, or in the case of one:many matching, a given
control subject can only be matched to a given treated subject once.
}
\references{
Pimentel, SD, Forrow, LV, Gellar, J, and J Li (2019). Optimal matching approaches in health policy evaluations under
rolling enrolment. \emph{Journal of the Royal Statistical Society Series A 183}(4), 1411-1435.
 https://doi.org/10.1111/rssa.12521
}
\keyword{enrollment}
\keyword{nonparametric}
\keyword{optimize}
\keyword{rolling}
