% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{matrix_to_set}
\alias{matrix_to_set}
\title{Convert GroupMatch match matrix to optmatch setID format}
\usage{
matrix_to_set(match)
}
\arguments{
\item{match}{GroupMatch output}
}
\value{
Optmatch setID vector
}
\description{
Convert GroupMatch match matrix to optmatch setID format
}
