% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{matrix_to_weights}
\alias{matrix_to_weights}
\title{Convert GroupMatch match matrix to matching weights attached to original dataframe}
\usage{
matrix_to_weights(mat, dat)
}
\arguments{
\item{mat}{GroupMatch match matrix}

\item{dat}{Dataset GroupMatch was performed on}
}
\value{
Dataset with matching weights
}
\description{
Convert GroupMatch match matrix to matching weights attached to original dataframe
}
