% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupmatch.R
\name{setTryRecovery}
\alias{setTryRecovery}
\title{(Internal) Sets up option to try recovery in \code{groupmatch}.}
\usage{
setTryRecovery()
}
\description{
(Internal) Sets up option to try recovery in \code{groupmatch}.
}
